Cleaned up library
