+++ libs/splatnet/src/index.ts
@@ -1,8 +1,7 @@
-export * from './lib/connect/nso-connect.model';
-export * from './lib/connect/nso-game.model';
+export { NsoGame, NsoGameCookie, IdToken, AccessToken } from './lib/connect/model/nso-connect.model';
 
 export * from './lib/nso-error.class';
+export * from './lib/nso-operation.class';
 export * from './lib/nso-app.class';
 export * from './lib/connect/nso-connector.class';
 export * from './lib/connect/nso-game-connector.class';
-export * from './lib/nso-operation.class';
