--- libs/splatnet/src/lib/connect/nso-game.model.ts
@@ -1,12 +1,0 @@
-export interface NsoGame {
-  id: number;
-  host: string;
-  name: string;
-  cookieName: string;
-}
-
-export interface NsoGameCookie {
-  fullHeader?: string;
-  value: string;
-  expires: number;
-}
