+++ libs/splatnet/src/lib/model/nso-app-args.model.ts
@@ -1,0 +1,5 @@
+export interface NsoAppArgs {
+  userAgent: string;
+  version?: string;
+  versionCheckInterval?: number;
+}
